# optimize and simplify DELTA_BYTE_ARRAY encoding